scripts/langfuse_loader/requirements.txt
langfuse==3.14.5
pandas>=1.5.0
pytest-dotenv
